Add Azure KeyVault capability

+++ demo-infrastructure/init-creds.sh
@@ -19,16 +19,50 @@ then
     export TEMPLATE_POLICYXML=/home/aceuser/ace-server/TEAJDBC.policyxml
 fi
 
-echo "policy ${TEMPLATE_POLICYXML} before"
-cat ${TEMPLATE_POLICYXML}
-sed -i "s/DATABASE_NAME/`cat /run/secrets/jdbc/databaseName`/g" ${TEMPLATE_POLICYXML}
-sed -i "s/SERVER_NAME/`cat /run/secrets/jdbc/serverName`/g" ${TEMPLATE_POLICYXML}
-sed -i "s/PORT_NUMBER/`cat /run/secrets/jdbc/portNumber`/g" ${TEMPLATE_POLICYXML}
 
-echo "policy ${TEMPLATE_POLICYXML} after"
-cat ${TEMPLATE_POLICYXML}
-cp ${TEMPLATE_POLICYXML} /home/aceuser/ace-server/run/PreProdPolicies/
+if [[ -e "/mnt/secrets-store/xmlCredentials" ]]
+then
+    # Azure keyvault
+    echo "policy ${TEMPLATE_POLICYXML} before Azure KV mods"
+    cat ${TEMPLATE_POLICYXML}
+    sed -i "s/DATABASE_NAME/${TEA_DATABASE_NAME}/g" ${TEMPLATE_POLICYXML}
+    sed -i "s/SERVER_NAME/${TEA_SERVER_NAME}/g" ${TEMPLATE_POLICYXML}
+    sed -i "s/PORT_NUMBER/${TEA_PORT_NUMBER}/g" ${TEMPLATE_POLICYXML}
+    
+    echo "policy ${TEMPLATE_POLICYXML} after"
+    cat ${TEMPLATE_POLICYXML}
+    cp ${TEMPLATE_POLICYXML} /home/aceuser/ace-server/run/PreProdPolicies/
 
-mqsisetdbparms -w /home/aceuser/ace-server -n jdbc::tea -u `cat /run/secrets/jdbc/USERID` -p `cat /run/secrets/jdbc/PASSWORD`
+    echo "" >> /home/aceuser/ace-server/server.conf.yaml
+    echo "ExternalCredentialsProviders:" >> /home/aceuser/ace-server/server.conf.yaml
+    echo "  TeaJDBCAzureKV:" >> /home/aceuser/ace-server/server.conf.yaml
+    echo "    loadAllCredentialsCommand: '/home/aceuser/ace-server/read-xml-creds.sh'" >> /home/aceuser/ace-server/server.conf.yaml
+    echo "    loadAllCredentialsFormat: 'xml'" >> /home/aceuser/ace-server/server.conf.yaml
+    echo ""  >> /home/aceuser/ace-server/server.conf.yaml
+fi
+
+# Add Hashicorp Vault section here
+
+
+
+if [[ -e "/home/aceuser/ace-server/run/PreProdPolicies/TEAJDBC.policyxml" ]]
+then
+    # Already completed
+    echo "Not reading kube secrets"
+else
+    # Original kube secrets approach
+    echo "policy ${TEMPLATE_POLICYXML} before"
+    cat ${TEMPLATE_POLICYXML}
+    sed -i "s/DATABASE_NAME/`cat /run/secrets/jdbc/databaseName`/g" ${TEMPLATE_POLICYXML}
+    sed -i "s/SERVER_NAME/`cat /run/secrets/jdbc/serverName`/g" ${TEMPLATE_POLICYXML}
+    sed -i "s/PORT_NUMBER/`cat /run/secrets/jdbc/portNumber`/g" ${TEMPLATE_POLICYXML}
+    
+    echo "policy ${TEMPLATE_POLICYXML} after"
+    cat ${TEMPLATE_POLICYXML}
+    cp ${TEMPLATE_POLICYXML} /home/aceuser/ace-server/run/PreProdPolicies/
+    
+    mqsisetdbparms -w /home/aceuser/ace-server -n jdbc::tea -u `cat /run/secrets/jdbc/USERID` -p `cat /run/secrets/jdbc/PASSWORD`
+fi
+
 
 sed -i "s/#policyProject: 'DefaultPolicies'/policyProject: 'PreProdPolicies'/g" /home/aceuser/ace-server/server.conf.yaml
